requirements.txt
pywinauto
asyncio
telethon
python-dotenv